Moved logic specific to transformations out of HapiHelper
