+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/ruleengine/transformation/custom/addSendingFacilityToMessageHeader.java
@@ -8,6 +8,8 @@ import gov.hhs.cdc.trustedintermediary.external.hapi.HapiHelper;
 import gov.hhs.cdc.trustedintermediary.wrappers.MetricMetadata;
 import java.util.Map;
 import org.hl7.fhir.r4.model.Bundle;
+import org.hl7.fhir.r4.model.Organization;
+import org.hl7.fhir.r4.model.Reference;
 
 public class addSendingFacilityToMessageHeader implements CustomFhirTransformation {
 
@@ -17,7 +19,10 @@ public class addSendingFacilityToMessageHeader implements CustomFhirTransformati
     @Override
     public void transform(FhirResource<?> resource, Map<String, String> args) {
         Bundle bundle = (Bundle) resource.getUnderlyingResource();
-        HapiHelper.addSendingFacilityToMessageHeader(bundle, args.get("name"));
+        var header = HapiHelper.findOrCreateMessageHeader(bundle);
+        var organization = new Organization();
+        organization.setName(args.get("name"));
+        header.setSender(new Reference(organization));
         metadata.put(bundle.getId(), EtorMetadataStep.CONTACT_SECTION_ADDED_TO_PATIENT);
     }
 }
