+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/ruleengine/transformation/custom/switchPlacerOrderAndGroupNumbers.java
@@ -8,6 +8,7 @@ import gov.hhs.cdc.trustedintermediary.external.hapi.HapiHelper;
 import gov.hhs.cdc.trustedintermediary.wrappers.MetricMetadata;
 import java.util.Map;
 import org.hl7.fhir.r4.model.Bundle;
+import org.hl7.fhir.r4.model.ServiceRequest;
 
 public class switchPlacerOrderAndGroupNumbers implements CustomFhirTransformation {
 
@@ -16,8 +17,37 @@ public class switchPlacerOrderAndGroupNumbers implements CustomFhirTransformatio
 
     @Override
     public void transform(FhirResource<?> resource, Map<String, String> args) {
+        //        Update ORC-2 with content from ORC-4 in the ORU result message.
+        //                Replace ORC-2.1 with content of ORC-4.1.
+        //                Replace ORC-2.2 with content of ORC-4.2
+        //        Effectively, we're swapping ORC-2 for ORC-4 and vice versa
+        //        OBR 2.1: identifier[0]. value
+        //        OBR 2.2: identifier.extension[1].extension[0].valueString
+        //        OBR 4.1: code.coding[0].code
+        //        OBR 4.2: code.coding[0].display
         Bundle bundle = (Bundle) resource.getUnderlyingResource();
-        HapiHelper.switchPlacerOrderAndGroupNumbers(bundle);
+        var serviceRequests = HapiHelper.resourcesInBundle(bundle, ServiceRequest.class);
+
+        serviceRequests.forEach(
+                serviceRequest -> {
+                    var twoPointOne = serviceRequest.getIdentifier().get(0);
+                    var twoPointTwo =
+                            serviceRequest
+                                    .getIdentifier()
+                                    .get(0)
+                                    .getExtensionByUrl(
+                                            "https://reportstream.cdc.gov/fhir/StructureDefinition/namespace-id");
+                    var fourPointOne = serviceRequest.getCode().getCoding().get(0);
+                    var fourPointTwo = serviceRequest.getCode().getCoding().get(0);
+
+                    var valueHolder = twoPointOne.getValue();
+                    twoPointOne.setValue(fourPointOne.getCode());
+                    fourPointOne.setCode(valueHolder);
+
+                    valueHolder = twoPointTwo.getValue().primitiveValue();
+                    twoPointTwo.setValue(fourPointTwo.getDisplayElement());
+                    fourPointTwo.setCode(valueHolder);
+                });
         metadata.put(bundle.getId(), EtorMetadataStep.SWITCH_PLACER_ORDER_AND_GROUP_NUMBERS);
     }
 }
